Separate MessageQueue and create Playlist inheriting from it
