+++ src/music_player.cpp
@@ -1,26 +1,9 @@
 #include <iostream>
 #include <thread>
-#include "music_player.hpp"
-#include "mp3.hpp"
 #include "fmod_errors.h"
 
-template <typename T>
-T MessageQueue<T>::pop()
-{
-    std::unique_lock lock{mtx_};
-    msg_available_.wait(lock, [this]{ return !queue_.empty(); });
-    auto msg = std::move(queue_.front());
-    queue_.pop();
-    return msg;
-}
-
-template <typename T>
-void MessageQueue<T>::push(T msg)
-{
-    std::lock_guard lock{mtx_};
-    queue_.emplace(std::move(msg));
-    msg_available_.notify_one();
-}
+#include "music_player.hpp"
+#include "mp3.hpp"
 
 void MusicPlayer::checkFmodOperation(const std::string &message, FMOD_RESULT result)
 {
@@ -97,7 +80,6 @@ void MusicPlayer::run()
             std::cout << e.what() << std::endl;
             std::cout << "The file you tried to play does not exist or it is not an audio file" << std::endl;
         }
-                
         sound->release();
         current_audio_.reset();
     }
