+++ src/music_player.hpp
@@ -2,30 +2,12 @@
 #define _MUSIC_PLAYER_HPP_
 
 #include <string>
-#include <queue>
-#include <mutex>
-#include <condition_variable>
 #include <memory>
 #include "fmod.hpp"
 
+#include "playlist.hpp"
 #include "audio.hpp"
 
-template <typename T>
-class MessageQueue 
-{
-    public:
-        void push(T msg);
-        T pop();
-        void printContent();
-    
-    private:
-        std::queue<T> queue_;
-        std::mutex mtx_;
-        std::condition_variable msg_available_;
-
-};
-
-//TODO Inherit from Threadable
 class MusicPlayer
 {
     public:
@@ -49,7 +31,8 @@ class MusicPlayer
         FMOD::Sound * currentsound_ {nullptr};
         FMOD::Channel* channel_ {nullptr};
 
-        MessageQueue<std::unique_ptr<Audio>> playlist_;
+        Playlist playlist_;
+
         std::unique_ptr<Audio> current_audio_;
         bool shutdown_{false};
